requirements.txt
opencv-python
mediapipe
pygrabber
pynput
numpy
pillow
tensorflow
scikit-learn
matplotlib